#430: make the whole culture surface int64 -- late-game overflow is real, not theoretical
Owner confirms culture overflow bites in late game, which is what the saturating
`< 0 ? MAX_INT` guards in the city getters were quietly absorbing. The previous
commit widened the city array; this finishes the surface, because a widened
store whose readers narrow again just moves the wrap.

- CvCity::calculateCulturePercent took the (now int64) total straight back into
  an `int` local -- the truncation relocating one line down. The operands stay
  wide and only the PERCENT, which is genuinely small, comes back as int.
- CvPlot::m_aiCulture is the same accumulating quantity one scope down: it takes
  the spread from city culture at x100 and only DECAYS proportionally, so it
  reaches equilibrium rather than being bounded. Storage, getCulture,
  countTotalCulture, countFriendlyCulture, setCulture and changeCulture all
  widen with it.

⚑ The per-PLAYER dimension is not incidental and is worth stating so it is not
"simplified" later: cultural OWNERSHIP is a contest, so calculateCulturePercent,
findHighestCulture and calculateCulturalOwner all ask how much culture a SPECIFIC
player has in that city or on that plot. A single total would say a place has
culture without saying whose. The dimension costs ~75 KB across a late-game map.

No save work: widening is soft now (save.md 8, the reader absorbs the narrower
stored form), so plot and city culture keep their tags and old saves widen in
place.

Assert build: no error names any culture symbol on the city/plot side; the
remaining `getCulture` errors are CvImprovementInfo/CvTerrainInfo info-getter
debt, a separate sweep.

+++ Sources/Engine/CvCity.cpp
@@ -10599,11 +10599,12 @@ PlayerTypes CvCity::findHighestCulture() const
 
 int CvCity::calculateCulturePercent(PlayerTypes eIndex) const
 {
-	const int iTotalCulture = countTotalCultureTimes100();
+	const int64_t iTotalCulture = countTotalCultureTimes100();
 
 	if (iTotalCulture > 0)
 	{
-		return getCultureTimes100(eIndex) * 100 / iTotalCulture;
+		// A PERCENT is the answer, so it stays int -- but the operands must not narrow on the way in.
+		return (int)(getCultureTimes100(eIndex) * 100 / iTotalCulture);
 	}
 	return 0;
 }
